django-react/django_react/requirements.txt
alabaster==0.7.12
appdirs==1.4.3
asgiref==3.2.7
astroid==2.2.5
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
distlib==0.3.0
Django==3.0.6
djangorestframework==3.11.0
docutils==0.16
filelock==3.0.12
idna==2.9
imagesize==1.2.0
isort==4.3.21
Jinja2==2.11.2
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.3
pbr==5.4.5
Pillow==7.1.2
Pygments==2.6.1
pylint==2.3.1
pyparsing==2.4.7
pytz==2020.1
requests==2.23.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==3.0.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
stevedore==1.32.0
typed-ast==1.4.0
urllib3==1.25.9
virtualenv==20.0.18
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wrapt==1.11.2
